fix: 온보딩 Fragment 내 데이터 처리 함수 Util 객체에서 관리
